BUGFIX: Add separator for multi dimensional usages
If you use more than one dimension, the dimension name is directly after the dimension value before. And that leads to a bad user experience as it is not good to read.

For instance, “country: pollanguage: pl”
With this change, we add a pipe and it will look like  that “country: pol | language: pl” in the usage table.

The comma as separator was already used if the dimension value has more than one item.